+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 ### Version 0.1.41
-- fix
+- fix iframe reference
 
 ### Version 0.1.40
 - Update display-card